pyproject.toml
[project]
name = "kazuha-invest-engine"
version = "0.1.0"
description = "Kazuha Invest 2.0: The Anticipatory Intelligence System"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckduckgo-search>=6.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "python-telegram-bot>=21.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
